LevelProvider: Refactor providers to drop lots of duplicated code
